더미데이터 위치 수정 및 빠진 CSS 추가
